Create README.md
When the customer buys some product then a bill is generated in which the product name and the total amount to be paid is mentioned along with the date and time of transaction. Then the store's inventory gets updated and the total sales file is also created.
When the store is purchasing new items then also the store's inventory gets updated accordingly.